+++ src/operators/mechanics/MechanicsNonlinearFEOperator.cc
@@ -227,7 +227,7 @@ namespace AMP {
       }//end r
 
       d_elementOutputVector.resize(3*numNodesInCurrElem);
-      for(int i = 0; i < d_elementOutputVector.size(); ++i) {
+      for(size_t i = 0; i < d_elementOutputVector.size(); ++i) {
         d_elementOutputVector[i] = 0.0; 
       }//end i
